DE VERSION 4.1/requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
numpy>=1.21.0
pillow>=9.0.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0 